Use pg_owner for dbt2 on driver and client nodes
This is for ease of benchmarking, to use operating system users that are
associated with the database owner and is thus able to gather statistics
per process with tools like pidstat without further permissions
configuration with the operating system.

With PostgreSQL, this is simply done by matching the default database
superuser to the operating system user.

This also remove extra vars and tasks that had previously set up a
separate dbt2 user.

The environment variable are now defined in .dbt2rc as opposed to the
shell's rc file.  This is anticipation for adding additional tests which
may have conflicting environment variable requirements.  So we will need
the user to source the environment file required for the test being run.

The operating system on the driver needs passwordless ssh to all of the
other systems under test in order to collect system statistics.  Update
the tasks such that the user on the driver creates a passwordless ssh
key and its public key is copied to all of the client and database
systems.